+++ Packages/AdvancedNewFile/AdvancedNewFile.py
@@ -4,6 +4,7 @@ import sublime_plugin
 import re
 import logging
 import errno
+import shutil
 
 SETTINGS = [
     "alias",
@@ -23,7 +24,12 @@ SETTINGS = [
     "auto_refresh_sidebar",
     "completion_type",
     "complete_single_entry",
-    "use_folder_name"
+    "use_folder_name",
+    "relative_from_current",
+    "default_extension",
+    "file_permissions",
+    "folder_permissions",
+    "rename_default"
 ]
 VIEW_NAME = "AdvancedNewFileCreation"
 WIN_ROOT_REGEX = r"[a-zA-Z]:(/|\\)"
@@ -38,13 +44,15 @@ logger = logging.getLogger()
 
 
 class AdvancedNewFileCommand(sublime_plugin.WindowCommand):
-    def run(self, is_python=False, initial_path=None):
+    def run(self, is_python=False, initial_path=None, rename=False, rename_file=None):
         PLATFORM = sublime.platform().lower()
         self.root = None
         self.alias_root = None
         self.top_level_split_char = ":"
         self.is_python = is_python
+        self.rename = rename
         self.view = self.window.active_view()
+        self.rename_filename = rename_file
 
         # Settings will be based on the view
         self.settings = get_settings(self.view)
@@ -59,10 +67,17 @@ class AdvancedNewFileCommand(sublime_plugin.WindowCommand):
         self.root, path = self.split_path(default_root)
 
         # Search for initial string
+
         if initial_path is not None:
             path = initial_path
         else:
-            path = self.settings.get("default_initial", "")
+            if rename:
+                path = self.settings.get("rename_default", "")
+                current_file = self.view.file_name()
+                current_file_name = os.path.basename(self.view.file_name()) if current_file else ""
+                path = path.replace("<current>", current_file_name)
+            if path == "":
+                path = self.settings.get("default_initial", "")
             if self.settings.get("use_cursor_text", False):
                 tmp = self.get_cursor_path()
                 if tmp != "":
@@ -139,10 +154,30 @@ class AdvancedNewFileCommand(sublime_plugin.WindowCommand):
                 if parts[1] != "":
                     path_list.append(parts[1])
                 path = self.top_level_split_char.join(path_list)
+            elif re.match(r"^/", path):
+                path_offset = 1
+                if PLATFORM == "windows":
+                    match = re.match(r"^/([a-zA-Z])/", path)
+                    if match:
+                        root = "%s:\\" % match.group(1)
+                        path_offset = 3
+                    else:
+                        root, _ = os.path.splitdrive(self.view.file_name())
+                        root += "\\"
+                else:
+                    root = "/"
+                path = path[path_offset:]
             # Parse if tilde used
             elif re.match(HOME_REGEX, path) and root == None:
                 root = os.path.expanduser("~")
                 path = path[2:]
+            elif re.match(r"^\.{1,2}[/\\]", path) and self.settings.get("relative_from_current", False):
+                path_index = 2
+                root = os.path.dirname(self.view.file_name())
+                if re.match(r"^\.{2}[/\\]", path):
+                    root = os.path.dirname(root)
+                    path_index = 3
+                path = path[path_index:]
 
             # Default
             if root == None:
@@ -155,6 +190,8 @@ class AdvancedNewFileCommand(sublime_plugin.WindowCommand):
                 root = root or self.window.folders()[folder_index]
         except IndexError:
             root = os.path.expanduser("~")
+
+
         return root, path
 
     def translate_alias(self, path):
@@ -223,7 +260,8 @@ class AdvancedNewFileCommand(sublime_plugin.WindowCommand):
     def update_filename_input(self, path_in):
         if self.settings.get("completion_type") == "windows":
             if "prev_text" in dir(self) and self.prev_text != path_in:
-                self.view.erase_status("AdvancedNewFile2")
+                if self.view is not None:
+                    self.view.erase_status("AdvancedNewFile2")
         if path_in.endswith("\t"):
             path_in = path_in.replace("\t", "")
             if self.settings.get("completion_type") == "windows":
@@ -233,13 +271,20 @@ class AdvancedNewFileCommand(sublime_plugin.WindowCommand):
 
         base, path = self.split_path(path_in)
 
-        creation_path = self.generate_creation_path(base, path)
+        creation_path = self.generate_creation_path(base, path, True)
         if self.show_path:
             if self.view != None:
-                self.view.set_status("AdvancedNewFile", "Creating file at %s " % \
-                    creation_path)
+                if self.rename:
+                    self.view.set_status("AdvancedNewFile", "Moving file to %s " % \
+                        creation_path)
+                else:
+                    self.view.set_status("AdvancedNewFile", "Creating file at %s " % \
+                        creation_path)
             else:
-                sublime.status_message("Creating file at %s" % creation_path)
+                if self.rename:
+                    sublime.status_message("Moving file to %s" % creation_path)
+                else:
+                    sublime.status_message("Creating file at %s" % creation_path)
         logger.debug("Creation path is '%s'" % creation_path)
 
     def generate_completion_list(self, path_in, each_list=False):
@@ -255,9 +300,10 @@ class AdvancedNewFileCommand(sublime_plugin.WindowCommand):
                 alias_list += self.generate_alias_auto_complete(filename)
                 alias_list += self.generate_project_auto_complete(filename)
         base, path = self.split_path(path_in)
-        directory, filename = os.path.split(path)
+        full_path = self.generate_creation_path(base, path)
 
-        directory = os.path.join(base, directory)
+        directory, filename = os.path.split(full_path)
+
         if os.path.isdir(directory):
             for d in os.listdir(directory):
                 full_path = os.path.join(directory, d)
@@ -318,10 +364,11 @@ class AdvancedNewFileCommand(sublime_plugin.WindowCommand):
 
         if len(self.completion_list) > 1:
             if first_token:
-                if self.completion_list[self.offset] in self.alias_list:
-                    self.view.set_status("AdvancedNewFile2", "Alias Completion")
-                elif self.completion_list[self.offset] in self.dir_list:
-                    self.view.set_status("AdvancedNewFile2", "Directory Completion")
+                if self.view is not None:
+                    if self.completion_list[self.offset] in self.alias_list:
+                        self.view.set_status("AdvancedNewFile2", "Alias Completion")
+                    elif self.completion_list[self.offset] in self.dir_list:
+                        self.view.set_status("AdvancedNewFile2", "Directory Completion")
             self.prev_text = new_content
         else:
             self.prev_text = None
@@ -389,7 +436,7 @@ class AdvancedNewFileCommand(sublime_plugin.WindowCommand):
         return compare_entry.startswith(compare_base)
 
 
-    def generate_creation_path(self, base, path):
+    def generate_creation_path(self, base, path, append_extension=False):
         if PLATFORM == "windows":
             if not re.match(WIN_ROOT_REGEX, base):
                 return base + self.top_level_split_char + path
@@ -397,7 +444,23 @@ class AdvancedNewFileCommand(sublime_plugin.WindowCommand):
             if not re.match(NIX_ROOT_REGEX, base):
                 return base + self.top_level_split_char + path
 
-        return os.path.abspath(os.path.join(base, path))
+        tokens = re.split(r"[/\\]", base) + re.split(r"[/\\]", path)
+        if tokens[0] == "":
+            tokens[0] = "/"
+        if PLATFORM == "windows":
+            tokens[0] = base[0:3]
+
+        full_path = os.path.abspath(os.path.join(*tokens))
+        if re.search(r"[/\\]$", path) or len(path) == 0:
+            full_path += os.path.sep
+        elif re.search(r"\.", tokens[-1]):
+            if re.search(r"\.$", tokens[-1]):
+                full_path += "."
+        elif append_extension:
+            filename = os.path.basename(full_path)
+            if not os.path.exists(full_path):
+                full_path += self.settings.get("default_extension", "")
+        return full_path
 
     def entered_filename(self, filename):
         # Check if valid root specified for windows.
@@ -410,7 +473,7 @@ class AdvancedNewFileCommand(sublime_plugin.WindowCommand):
                     return
 
         base, path = self.split_path(filename)
-        file_path = os.path.join(base, path)
+        file_path = self.generate_creation_path(base, path, True)
         # Check for invalid alias specified.
         if self.top_level_split_char in filename and \
             not (PLATFORM == "windows" and re.match(WIN_ROOT_REGEX, base)) and \
@@ -431,13 +494,65 @@ class AdvancedNewFileCommand(sublime_plugin.WindowCommand):
                     sublime.error_message("Cannot create '" + file_path + "'. See console for details")
                     logger.error("Exception: %s '%s'" % (e.strerror, e.filename))
             if attempt_open:
-                if os.path.isdir(file_path):
-                    if not re.search(r"(/|\\)$", file_path):
-                        sublime.error_message("Cannot open view for '" + file_path + "'. It is a directory. ")
+                if self.rename:
+                    self.rename_file(file_path)
                 else:
-                    self.window.open_file(file_path)
+                    self.open_file(file_path)
+
         self.clear()
         self.refresh_sidebar()
+
+    def open_file(self, file_path):
+        new_view = None
+        if os.path.isdir(file_path):
+            if not re.search(r"(/|\\)$", file_path):
+                sublime.error_message("Cannot open view for '" + file_path + "'. It is a directory. ")
+        else:
+            new_view = self.window.open_file(file_path)
+        return new_view
+
+    def rename_file(self, file_path):
+        if os.path.isdir(file_path):
+            if not re.search(r"(/|\\)$", file_path):
+                sublime.error_message("Cannot open view for '" + file_path + "'. It is a directory. ")
+        else:
+            window = self.window
+            if self.rename_filename:
+                shutil.move(self.rename_filename, file_path)
+                file_view = self.find_open_file(self.rename_filename)
+                if file_view is not None:
+                    window.focus_view(file_view)
+                    window.run_command("close")
+                    self.open_file(file_path)
+
+            elif self.view:
+                if self.view.file_name():
+                    self.view.run_command("save")
+                    window.focus_view(self.view)
+                    window.run_command("close")
+                    shutil.move(self.view.file_name(), file_path)
+                else:
+                    content = self.view.substr(sublime.Region(0, self.view.size()))
+                    self.view.set_scratch(True)
+                    self.view.run_command("close")
+                    window.focus_view(self.view)
+                    window.run_command("close")
+                    with open(file_path, "w") as file_obj:
+                        file_obj.write(content)
+                self.open_file(file_path)
+            else:
+                sublime.error_message("Unable to move file. No file to move.")
+
+    def find_open_file(self, file_name):
+        window = self.window
+        if IS_ST3:
+            return window.find_open_file(file_name)
+        else:
+            for view in window.views():
+                view_name = view.file_name()
+                if view_name != "" and view_name == file_name:
+                    return view
+        return None
 
     def refresh_sidebar(self):
         if self.settings.get("auto_refresh_sidebar"):
@@ -455,23 +570,36 @@ class AdvancedNewFileCommand(sublime_plugin.WindowCommand):
         base, filename = os.path.split(filename)
         self.create_folder(base)
         if filename != "":
-            open(os.path.join(base, filename), "a").close()
+            creation_path = os.path.join(base, filename)
+            open(creation_path, "a").close()
+            if self.settings.get("file_permissions", "") != "":
+                file_permissions = self.settings.get("file_permissions", "")
+                os.chmod(creation_path, int(file_permissions, 8))
 
     def create_folder(self, path):
         init_list = []
-        if self.is_python:
-            temp_path = path
-            while not os.path.exists(temp_path):
-                init_list.append(temp_path)
-                temp_path = os.path.dirname(temp_path)
+        temp_path = path
+        while not os.path.exists(temp_path):
+            init_list.append(temp_path)
+            temp_path = os.path.dirname(temp_path)
         try:
-            os.makedirs(path)
+            if not os.path.exists(path):
+                os.makedirs(path)
         except OSError as ex:
             if ex.errno != errno.EEXIST:
                 raise
 
+        file_permissions = self.settings.get("file_permissions", "")
+        folder_permissions = self.settings.get("folder_permissions", "")
         for entry in init_list:
-            open(os.path.join(entry, '__init__.py'), 'a').close()
+            if self.is_python:
+                creation_path = os.path.join(entry, '__init__.py')
+                open(creation_path, 'a').close()
+                if  file_permissions != "":
+                    os.chmod(creation_path, int(file_permissions, 8))
+            if  folder_permissions != "":
+                os.chmod(entry, int(folder_permissions, 8))
+
 
     def get_cursor_path(self):
         if self.view == None:
@@ -509,6 +637,12 @@ class AdvancedNewFileAtCommand(sublime_plugin.WindowCommand):
         settings = sublime.load_settings("AdvancedNewFile.sublime-settings")
         return settings.get("show_sidebar_menu", False) and len(dirs) == 1
 
+class AdvancedNewFileRenameAtCommand(sublime_plugin.WindowCommand):
+    def run(self, files):
+        self.window.run_command("advanced_new_file", {"rename": True, "rename_file": files[0]})
+
+    def is_visible(self, files):
+        return len(files) == 1
 
 def get_settings(view):
     settings = sublime.load_settings("AdvancedNewFile.sublime-settings")